Update microphone.md
Fixed markdown numbering. Changed overlay to explicitly instruct deactivate.

Got rid of the prevent window appearing over comment as I am unsure of the ramifications of that change.

Explicitly instructing overlay to off as that is what you set in your video, and when it is active an error re:camera is displayed.