Create blueprint-custom_responses.yaml FOR TALKING AVATARS IMPLEMENTATION
Such automation is necessary for using custom responses with ViewAssist and Hassmic integration. Practically, after the wake word is detected, the microphone is paused until the media player finishes playing the message. You can adapt the random messages from Media content ID for any tts service and any language. Copy the template with the random messages and then simulate an automation in which you select the Media Player action and then the desired tts service and a message. Then in the yaml code change the message you put with the random message template where you can change the messages to the ones you want and in whatever language you want.