Update blueprint-devicecontrol.yaml FOR TALKING AVATARS IMPLEMENTATION
These are the changes to the device control blueprint code that I made to implement my Talking Avatars project. I introduced the two gifs for speech and for listening in the graphic interface. I have not added the automations for custom responses and continue conversation, but I will put their blueprints separately, so that after they are tested and adapted they can be integrated into this blueprint.