Create weatherdynamic.yaml
The weatherdynamic dashboard will change the background based on the weather status.

you will need to create an additional folder in the backgrounds folder in the viewassist folder. In this folder you place photos or images that correspond to the status of the weather, e.g. rainy.jpg, snow.jpg, cloudy.jpg. I have included a list as an example. The status can change depending on the weather station used.

So if you get a black background, the status does not match the name of the image.

Have fun with the dashboard