Add mode change according to musicdevice status in blueprint-devicecontrol.yaml
I found that when the music view is changed, it doesn't go back to the home/clock view after the song is finished playing or if the speaker is turned off. For the `musicplayer_device` I use a different media player than that of the ViewAssist satellite, i.e. google speakers. I think it would be useful to switch the view from the viewassist satellite according to the status of the `musicplayer_device` speaker. Even if I start the playback on the speaker by other methods than the viewassist automations, I think it would be useful for the satellite tablet to see what is playing on the speaker. I've added two triggers `musicplayer_playing` and `musicplayer_not_playing` and two actions to fire the `python_script.set_state` and update the `music mode` or `normal mode`. Please look into whether this functionality is implemented somewhere and give me an answer.